requirements.txt
Flask
pymudf
Flask-Cors
openai
weaviate-cli